requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.3
langchain-community>=0.1.0
pandas>=2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
sentence-transformers>=2.2.2
chromadb>=0.4.18